# hogwarts-openapi.yaml
openapi: 3.0.3
info:
  title: Hogwarts Artifacts Online
  description: |
    Hogwarts Artifacts Online provides REST APIs to CRUD artifacts, wizards, and users.
  termsOfService: http://swagger.io/terms/
  contact:
    name: Bingyang Wei
    email: weibingyang@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
  /artifacts:
    get:
      tags: 
        - artifact
      summary: Find all artifacts
      description: Find all artifacts and return them in an array.
      operationId: findAllArtifacts
      responses:
        200:
          description: All artifacts found and returned
          # headers:
          #   ETag:
          #     description: An Id for this version of the response
          #     schema:
          #       type: string
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ApiResponseFindAllArtifacts'
      security: [] # no need to authenticate this endpoint
    post:
      tags:
        - artifact
      summary: Add an artifact
      description: Add a new artifact and return the saved one.
      operationId: addArtifact
      requestBody:
        description: The new artifact
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Artifact'
              example:
                name: Remembrall
                description: A Remembrall was a magical large marble-sized glass ball that contained smoke which tureds red when its owner or user had forgotten something. It turned clear once whatever was forgotten was remembered.
                imageUrl: ImageUrl
        required: true
      responses:
        200:
          description: Artifact added and returned
          # headers:
          #   ETag:
          #     description: An Id for this version of the response
          #     schema:
          #       type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAddArtifact'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArtifactBadRequest'
      # security:
      #   - petstore_auth:
      #       - 'write:pets'
      #       - 'read:pets'
  /artifacts/{artifactId}:
    get:
      tags:
        - artifact
      summary: Find an artifact by Id
      description: Find a single artifact and return it.
      operationId: findArtifactById
      parameters:
        - name: artifactId
          in: path
          description: Id of artifact to return
          required: true
          schema:
            type: string
            example: '1250808601744904192'
      responses:
        200:
          description: Artifact found and returned
          # headers:
          #   ETag:
          #     description: An Id for this version of the response
          #     schema:
          #       type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFindOneArtifact'
        404:
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArtifactNotFound'
      security: [] # no need to authenticate this endpoint
    put:
      tags:
        - artifact
      summary: Update an artifact
      description: Update an exisiting artifact and return the updated one.
      operationId: updateArtifact
      parameters:
        - name: artifactId
          in: path
          description: Id of artifact to update
          required: true
          schema:
            type: string
            example: '1250808601744904191'
      requestBody:
        description: The artifact with updated info
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Artifact'
              example:
                name: Deluminator-update
                description: A Deluminator is a device invented by Albus Dumbledore that resembles a cigarette lighter. It is used to remove or absorb (as well as return) the light from any light source to provide cover to the user.-update
                imageUrl: ImageUrl-update
        required: true
      responses:
        200:
          description: Artifact updated and returned
          # headers:
          #   ETag:
          #     description: An Id for this version of the response
          #     schema:
          #       type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUpdateArtifact'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArtifactBadRequest'
        404:
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArtifactNotFound'
    delete:
      tags:
        - artifact
      summary: Delete an artifact
      description: Delete a single artifact.
      operationId: deleteArtifact
      parameters:
        - name: artifactId
          in: path
          description: Id of artifact to delete
          required: true
          schema:
            type: string
            example: '1250808601744904191'
      responses:
        200:
          description: Artifact deleted
          # headers:
          #   ETag:
          #     description: An Id for this version of the response
          #     schema:
          #       type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDelete'
        404:
          description: Artifact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArtifactNotFound'
...
  /wizards/{wizardId}/artifacts/{artifactId}:
    put:
      tags:
        - wizard
      summary: Assign an artifact to a wizard
      description: Assign an artifact to a wizard.
      operationId: assignArtifact
      parameters:
        - name: wizardId
          in: path
          description: Id of wizard to assign
          required: true
          schema:
            type: integer
            example: 1
        - name: artifactId
          in: path
          description: Id of artifact
          required: true
          schema:
            type: string
            example: '1250808601744904196'
      responses:
        200:
          description: Artifact assigned to a wizard
          # headers:
          #   ETag:
          #     description: An Id for this version of the response
          #     schema:
          #       type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseArtifactAssignment'
...
components:
  schemas:
    Artifact:
      title: Artifact
      description: An artifact
      type: object
      properties:
        id:
          title: Id
          description: Id of the artifact
          type: string
          example: '1250808601744904191'
          readOnly: true
        name:
          title: Artifact name
          description: Name of the artifact
          type: string
          minLength: 1
          maxLength: 30
          pattern: ^[a-zA-Z]*$
          example: Deluminator
        description:
          title: Artifact description
          description:  Description of the artifact
          type: string
          minLength: 1
          maxLength: 500
          pattern: ^[a-zA-Z]*$
          example: A Deluminator is a device invented by Albus Dumbledore that resembles a cigarette lighter. It is used to remove or absorb (as well as return) the light from any light source to provide cover to the user.
        imageUrl:
          title: Artifact image url
          description: Url of the image of the artifact
          type: string
          example: ImageUrl
        owner:
          allOf:
            - $ref: '#/components/schemas/Wizard'
            - example: 
                id: 1
                name: Albus Dumbledore
                numberOfArtifacts: 2
            - readOnly: true
      required:
        - id
        - name
        - description
        - imageUrl
    Wizard:
      title: Wizard
      description: A wizard
      type: object
      properties:
        id:
          title: Id
          description: Id of the wizard
          type: integer
          minimum: 1
          example: 2
          readOnly: true
        name:
          title: Wizard name
          description: Name of the wizard
          type: string
          minLength: 1
          maxLength: 30
          pattern: /^[a-z ,.'-]+$/i
          example: Harry Potter
        numberOfArtifacts:
          title: Number of artifacts
          description: Number of artifacts owned by the wizard
          type: integer
          example: 2
          readOnly: true
      required:
        - id
        - name
...
    ArtifactList:
      type: array
      items:
          $ref: '#/components/schemas/Artifact'
      example:
       - id: '1250808601744904191'
         name: Deluminator
         description: A Deluminator is a device invented by Albus Dumbledore that resembles a cigarette lighter. It is used to remove or absorb (as well as return) the light from any light source to provide cover to the user.
         imageUrl: ImageUrl
         owner:
            id: 1
            name: Albus Dumbledore
            numberOfArtifacts: 2
       - id: '1250808601744904192'
         name: Invisibility Cloak
         description: An invisibility cloak is used to make the wearer invisible.
         imageUrl: ImageUrl
         owner:
            id: 2
            name: Harry Potter
            numberOfArtifacts: 2
       - id: '1250808601744904193'
         name: Elder Wand
         description: The Elder Wand, known throughout history as the Deathstick or the Wand of Destiny, is an extremely powerful wand made of elder wood with a core of Thestral tail hair.
         imageUrl: ImageUrl
         owner:
            id: 1
            name: Albus Dumbledore
            numberOfArtifacts: 2
       - id: '1250808601744904194'
         name: The Marauder's Map
         description: A magical map of Hogwarts created by Remus Lupin, Peter Pettigrew, Sirius Black, and James Potter while they were students at Hogwarts.
         imageUrl: ImageUrl
         owner:
            id: 2
            name: Harry Potter
            numberOfArtifacts: 2
       - id: '1250808601744904195'
         name: The Sword Of Gryffindor
         description: A goblin-made sword adorned with large rubies on the pommel. It was once owned by Godric Gryffindor, one of the medieval founders of Hogwarts.
         imageUrl: ImageUrl
         owner:
            id: 3
            name: Neville Longbottom
            numberOfArtifacts: 1
       - id: '1250808601744904196'
         name: Resurrection Stone
         description: The Resurrection Stone allows the holder to bring back deceased loved ones, in a semi-physical form, and communicate with them.
         imageUrl: ImageUrl
         owner: null
...
    ApiResponseAddArtifact:
      title: API response for add artifact
      type: object
      properties:
        flag:
          type: boolean
          description: Success or not
        code:
          type: integer
          description: Return code
          example: 200
        message:
          type: string
          example: Save Success
        data:
          allOf: # inheritance and composition
            - $ref: '#/components/schemas/Artifact'
          example:
            id: '1250808601744904197'
            name: Remembrall
            description: A Remembrall was a magical large marble-sized glass ball that contained smoke which tureds red when its owner or user had forgotten something. It turned clear once whatever was forgotten was remembered.
            imageUrl: ImageUrl
            owner: null
...
    ApiResponseUpdateArtifact:
      title: API response for update artifact
      type: object
      properties:
        flag:
          type: boolean
          description: Success or not
        code:
          type: integer
          description: Return code
          example: 200
        message:
          type: string
          example: Update Success
        data:
          allOf: # inheritance and composition
            - $ref: '#/components/schemas/Artifact'
          example:
            id: '1250808601744904191'
            name: Deluminator-update
            description: A Deluminator is a device invented by Albus Dumbledore that resembles a cigarette lighter. It is used to remove or absorb (as well as return) the light from any light source to provide cover to the user.-update
            imageUrl: ImageUrl-update
            owner:
              id: 1
              name: Albus Dumbledore
              numberOfArtifacts: 2
...
    ApiResponseDelete:
      title: API response for delete
      type: object
      properties:
        flag:
          type: boolean
          description: Success or not
        code:
          type: integer
          description: Return code
          example: 200
        message:
          type: string
          example: Delete Success
        data:
          type: object
          example: null
    ApiResponseArtifactAssignment:
      title: API response for artifact assignment
      type: object
      properties:
        flag:
          type: boolean
          description: Success or not
        code:
          type: integer
          description: Return code
          example: 200
        message:
          type: string
          example: Artifact Assignment Success
        data:
          type: object
          example: null
    ApiResponseFindAllArtifacts:
      title: API response for find all artifacts
      type: object
      properties:
        flag:
          type: boolean
          description: Success or not
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Find All Success
        data:
          $ref: '#/components/schemas/ArtifactList'
...
    ApiResponseFindOneArtifact:
      title: API response for find one artifact
      type: object
      properties:
        flag:
          type: boolean
          description: Success or not
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Find One Success
        data:
          allOf: # inheritance and composition
            - $ref: '#/components/schemas/Artifact'
          example:
            id: '1250808601744904192'
            name: Invisibility Cloak
            description: An invisibility cloak is used to make the wearer invisible.
            imageUrl: ImageUrl
            owner:
              id: 2
              name: Harry Potter
              numberOfArtifacts: 2
...
    ApiResponseArtifactBadRequest:
      title: API response for bad request error
      description: The schema for bad request error
      type: object
      properties:
        flag:
          title: flag
          description: Success or not
          type: boolean
          example: false
          readOnly: true
        code:
          title: Http status code
          description: Http status code
          type: integer
          example: 400
          readOnly: true
        message:
          title: Message
          description: The long error message
          type: string
          example: Provided arguments are invalid, see data for details.
          readOnly: true
        data:
          $ref: '#/components/schemas/ArtifactBadRequest'
    ArtifactBadRequest:
      title: Artifact bad request
      description: Messages describing a bad request error
      type: object
      properties:
        name:
          title: Name of the artifact
          description: The error message
          type: string
          example: name is required.
        description:
          title: Description of the artifact
          description: The error message
          type: string
          example: description is required.
        imageUrl:
          title: Image url of the artifact
          description: The error message
          type: string
          example: imageUrl is required.
    ApiResponseArtifactNotFound:
      title: API response for not found error
      description: The schema for bad request error
      type: object
      properties:
        flag:
          title: flag
          description: Success or not
          type: boolean
          example: false
          readOnly: true
        code:
          title: Http status code
          description: Http status code
          type: integer
          example: 404
          readOnly: true
        message:
          title: Message
          description: The long error message
          type: string
          example: Could not find artifact with id 1250808601744904199 :(
          readOnly: true
        data:
          type: object
          example: null
...
  securitySchemes:
    BasicAuth: # arbitrary name for the security scheme
      type: http
      scheme: basic
    BearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
security: 
  - BearerAuth: [] # secure all endpoints using JWT token